fix cxx warnings

+++ src/t8_default/t8_default_tet.c
@@ -115,7 +115,7 @@ static void
 t8_default_tet_set_linear_id (t8_element_t * elem, int level, uint64_t id)
 {
   T8_ASSERT (0 <= level && level <= T8_DTET_MAXLEVEL);
-  T8_ASSERT (0 <= id && id < 1 << 3 * level);
+  T8_ASSERT (0 <= id && id < (uint64_t) 1 << 3 * level);
 
   t8_dtet_init_linear_id ((t8_default_tet_t *) elem, id, level);
 }
