+++ src/t8_default/t8_dtri_bits.c
@@ -645,7 +645,7 @@ t8_dtri_init_linear_id (t8_dtri_t * t, uint64_t id, int level)
   t8_dtri_cube_id_t   cid;
   t8_dtri_type_t      type;
 
-  T8_ASSERT (0 <= id && id <= 1 << (T8_DTRI_DIM * level));
+  T8_ASSERT (0 <= id && id <= (u_int64_t) 1 << (T8_DTRI_DIM * level));
 
   t->level = level;
   t->x = 0;
